+++ core/server/services/channels/index.js
@@ -4,8 +4,10 @@
  * The channel service is responsible for:
  * - maintaining the config of available Channels
  * - building out the logic of how an individual Channel works
+ * - providing a top level router as an entry point
+ *
+ * Everything else exposed via the Channel object
  */
 
-module.exports.channelRouter = require('./router');
-// A mechanism to load the current channels
-module.exports.load = require('./loader').list;
+// Exports the top-level router
+module.exports = require('./parent-router');
