+++ core/server/services/channels/parent-router.js
@@ -1,5 +1,6 @@
-var ParentRouter = require('./ParentRouter'),
-    channelService = require('../channels');
+var ParentRouter = require('../route').ParentRouter,
+    loader = require('./loader'),
+    channelRouter = require('./router');
 
 /**
  * Channels Router
@@ -9,9 +10,9 @@ var ParentRouter = require('./ParentRouter'),
 module.exports.router = function channelsRouter() {
     var channelsRouter = new ParentRouter('channels');
 
-    channelService.load().forEach(function (channel) {
+    loader.list().forEach(function (channel) {
         // Create a new channelRouter, and mount it onto the parent at the correct route
-        channelsRouter.mountRouter(channel.route, channelService.channelRouter(channel));
+        channelsRouter.mountRouter(channel.route, channelRouter(channel));
     });
 
     return channelsRouter.router();
