+++ core/server/services/route/index.js
@@ -16,7 +16,11 @@
  * These both happen prior to the routeService router being mounted
  */
 
+// This is the main router, that gets mounted in the express app in /site
+module.exports.siteRouter = require('./site-router');
+
 // We expose this via the App Proxy, so that Apps can register routes
 module.exports.appRouter = require('./app-router');
-// This is the main router, that gets mounted in the express app in /site
-module.exports.router = require('./site-router');
+
+// Classes for other parts of Ghost to extend
+module.exports.ParentRouter = require('./ParentRouter');
